[Client] Fix LogFetcher doesn't update bucket metadata when receive NotLeaderOrFollowerException in bucket level
