+++ fluss-client/src/main/java/org/apache/fluss/client/table/scanner/log/LogFetcher.java
@@ -26,8 +26,10 @@ import org.apache.fluss.client.table.scanner.ScanRecord;
 import org.apache.fluss.cluster.BucketLocation;
 import org.apache.fluss.config.ConfigOptions;
 import org.apache.fluss.config.Configuration;
+import org.apache.fluss.exception.ApiException;
 import org.apache.fluss.exception.InvalidMetadataException;
 import org.apache.fluss.exception.LeaderNotAvailableException;
+import org.apache.fluss.exception.PartitionNotExistException;
 import org.apache.fluss.fs.FsPath;
 import org.apache.fluss.metadata.PhysicalTablePath;
 import org.apache.fluss.metadata.SchemaGetter;
@@ -48,6 +50,7 @@ import org.apache.fluss.rpc.messages.PbFetchLogReqForBucket;
 import org.apache.fluss.rpc.messages.PbFetchLogReqForTable;
 import org.apache.fluss.rpc.messages.PbFetchLogRespForBucket;
 import org.apache.fluss.rpc.messages.PbFetchLogRespForTable;
+import org.apache.fluss.rpc.protocol.ApiError;
 import org.apache.fluss.rpc.protocol.Errors;
 import org.apache.fluss.utils.IOUtils;
 import org.apache.fluss.utils.Projection;
@@ -217,14 +220,27 @@ public class LogFetcher implements Closeable {
             }
         }
 
-        if (isPartitioned && !partitionIds.isEmpty()) {
-            metadataUpdater.updateMetadata(Collections.singleton(tablePath), null, partitionIds);
-        } else if (needUpdate) {
-            metadataUpdater.updateTableOrPartitionMetadata(tablePath, null);
+        try {
+            if (isPartitioned && !partitionIds.isEmpty()) {
+                metadataUpdater.updateMetadata(
+                        Collections.singleton(tablePath), null, partitionIds);
+            } else if (needUpdate) {
+                metadataUpdater.updateTableOrPartitionMetadata(tablePath, null);
+            }
+        } catch (Exception e) {
+            if (e instanceof PartitionNotExistException) {
+                // ignore this exception, this is probably happen because the partition is deleted.
+                // The fetcher can also work fine. The caller like flink can remove the partition
+                // from fetch list when receive exception.
+                LOG.warn("Receive PartitionNotExistException when update metadata, ignore it", e);
+            } else {
+                throw e;
+            }
         }
     }
 
-    private void sendFetchRequest(int destination, FetchLogRequest fetchLogRequest) {
+    @VisibleForTesting
+    void sendFetchRequest(int destination, FetchLogRequest fetchLogRequest) {
         TableOrPartitions tableOrPartitionsInFetchRequest =
                 getTableOrPartitionsInFetchRequest(fetchLogRequest);
         // TODO cache the tablet server gateway.
@@ -345,6 +361,14 @@ public class LogFetcher implements Closeable {
                                     respForBucket.getBucketId());
                     FetchLogResultForBucket fetchResultForBucket =
                             getFetchLogResultForBucket(tb, tablePath, respForBucket);
+
+                    // if error code is not NONE, it means the fetch log request failed, we need to
+                    // clear table bucket meta for InvalidMetadataException.
+                    if (fetchResultForBucket.getErrorCode() != Errors.NONE.code()) {
+                        ApiError error = ApiError.fromErrorMessage(respForBucket);
+                        handleFetchLogExceptionForBucket(tb, destination, error);
+                    }
+
                     Long fetchOffset = logScannerStatus.getBucketOffset(tb);
                     // if the offset is null, it means the bucket has been unsubscribed,
                     // we just set a Long.MAX_VALUE as the next fetch offset
@@ -387,6 +411,29 @@ public class LogFetcher implements Closeable {
         }
     }
 
+    private void handleFetchLogExceptionForBucket(TableBucket tb, int destination, ApiError error) {
+        ApiException exception = error.error().exception();
+        LOG.error("Failed to fetch log from node {} for bucket {}", destination, tb, exception);
+        if (exception instanceof InvalidMetadataException) {
+            LOG.warn(
+                    "Invalid metadata error in fetch log request. "
+                            + "Going to request metadata update.",
+                    exception);
+            long tableId = tb.getTableId();
+            TableOrPartitions tableOrPartitions;
+            if (tb.getPartitionId() == null) {
+                tableOrPartitions = new TableOrPartitions(Collections.singleton(tableId), null);
+            } else {
+                tableOrPartitions =
+                        new TableOrPartitions(
+                                null,
+                                Collections.singleton(
+                                        new TablePartition(tableId, tb.getPartitionId())));
+            }
+            invalidTableOrPartitions(tableOrPartitions);
+        }
+    }
+
     private void pendRemoteFetches(
             RemoteLogFetchInfo remoteLogFetchInfo, long firstFetchOffset, long highWatermark) {
         checkNotNull(remoteLogFetchInfo);
@@ -417,7 +464,8 @@ public class LogFetcher implements Closeable {
         }
     }
 
-    private Map<Integer, FetchLogRequest> prepareFetchLogRequests() {
+    @VisibleForTesting
+    Map<Integer, FetchLogRequest> prepareFetchLogRequests() {
         Map<Integer, List<PbFetchLogReqForBucket>> fetchLogReqForBuckets = new HashMap<>();
         int readyForFetchCount = 0;
         Long tableId = null;
